Updated Experimental Rig
Changed the Pupil System from a pure External Texture to a Procedural System, providing for more control and a better look.

The way the pupils are scaled also was changed. The Pupil Height is now a property option in the Face Properties, though scaling the eyes by normal Means is still possible.